+++ src/Commands.ts
@@ -4,6 +4,10 @@
 import * as vscode from 'vscode';
 import * as cp from 'child_process';
 import { updatePathIndicator } from './StatusBar';
+import os = require('os');
+
+// Fix for weird unicode characters in terminal output readouts
+const outputReplace:any = /[^a-zA-Z0-9!?.:;\n{}\[\] ]/g;	
 
 function registerCommandNice(context: vscode.ExtensionContext, commandId: string, run: (...args: any[]) => void): void {
     context.subscriptions.push(vscode.commands.registerCommand(commandId, run));
@@ -13,6 +17,9 @@ export default function registerCommands(context: vscode.ExtensionContext) {
 
     // Get Praat path
     var praatPath:string = context.globalState.get('praatpath','');
+
+    // Create an output window to emulate Praat's
+	let praatOut = vscode.window.createOutputChannel("Praat InfoWindow");
 
     // Command for defining Praat path
     registerCommandNice(context, 'praatvscode.definePath', async() => {
@@ -39,7 +46,7 @@ export default function registerCommands(context: vscode.ExtensionContext) {
         // Push to global context
         await context.globalState.update('praatpath', promptPath);
         // Update status bar
-        // updatePathIndicator();
+        updatePathIndicator(context);
     });
 
     // Command for calling Praat path
@@ -48,6 +55,147 @@ export default function registerCommands(context: vscode.ExtensionContext) {
 			vscode.window.showErrorMessage('You have not set a path for the Praat executable yet. Run the command "Define a path for the Praat executable" to set a path.');
 		} else {
 			vscode.window.showInformationMessage('The current path for the Praat executable is: '+ praatPath);
+		}
+	});
+
+    // Command for opening a script in Praat
+    registerCommandNice(context, 'praatvscode.openInPraat', () => {
+		if (praatPath === undefined) {
+			vscode.window.showErrorMessage('You have not set a path for the Praat executable yet. Run the command "Define a path for the Praat executable" to set a path.');
+		} else if (vscode.window.activeTextEditor?.document.languageId !== "praat") {
+			vscode.window.showErrorMessage('The active selection is not a Praat script!');
+		} else {
+			if (vscode.window.activeTextEditor?.document.uri.fsPath !== undefined) {
+				if (os.type() === 'Windows_NT') {
+					cp.exec(praatPath +'\\Praat.exe --open ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'\\Praat.exe --open ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+					});
+				} else if (os.type() === 'Linux') {
+					cp.exec(praatPath +'/praat --open ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'/praat --open ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+					});
+				} else if (os.type() === 'Darwin') {
+					cp.exec(praatPath +'/Praat.app/Contents/MacOS/Praat --open ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'/Praat.app/Contents/MacOS/Praat --open ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+					});
+				} else {
+					vscode.window.showInformationMessage('Fatal error! Trouble recognizing OS.');
+				}
+			} else {
+				vscode.window.showErrorMessage('You do not have a Praat script open on VSCode!');
+			}
+		}
+	});
+
+    // Command for running a script in Praat
+	registerCommandNice(context, 'praatvscode.runInPraat', () => {
+		if (praatPath === undefined) {
+			vscode.window.showErrorMessage('You have not set a path for the Praat executable yet. Run the command "Define a path for the Praat executable" to set a path.');
+		} else if (vscode.window.activeTextEditor?.document.languageId !== "praat") {
+			vscode.window.showErrorMessage('The active selection is not a Praat script!');
+		} else {
+			if (vscode.window.activeTextEditor?.document.uri.fsPath !== undefined) {
+				if (os.type() === 'Windows_NT') {
+					cp.exec(praatPath +'\\Praat.exe --send ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'\\Praat.exe --send ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+					});
+				} else if (os.type() === 'Linux') {
+					cp.exec(praatPath +'/praat --send ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'/praat --send ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+					});
+				} else if (os.type() === 'Darwin') {
+					cp.exec(praatPath +'/Praat.app/Contents/MacOS/Praat --send ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'/Praat.app/Contents/MacOS/Praat --send ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+					});
+				} else {
+					vscode.window.showInformationMessage('Fatal error! Trouble recognizing OS.');
+				}
+			} else {
+				vscode.window.showErrorMessage('You do not have a Praat script open on VSCode!');
+			}
+		}
+	});
+
+    // Command for running a script in the background
+    registerCommandNice(context, 'praatvscode.runScript', () => {
+		if (praatPath === undefined) {
+			vscode.window.showErrorMessage('You have not set a path for the Praat executable yet. Run the command "Define a path for the Praat executable" to set a path.');
+		} else if (vscode.window.activeTextEditor?.document.languageId !== "praat") {
+			vscode.window.showErrorMessage('The active selection is not a Praat script!');
+		} else {
+			if (vscode.window.activeTextEditor?.document.uri.fsPath !== undefined) {
+				if (os.type() === 'Windows_NT') { // Gate check
+					cp.exec(praatPath +'\\Praat.exe --run ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'\\Praat.exe --run ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+						praatOut.show();
+						if (error) {
+							praatOut.appendLine('The script did not run.\nPraat sent:\n' + error);
+						} else {
+							praatOut.appendLine('The script ran nominally.\nPraat sent:\n' + stdout.replace(outputReplace, "") + ' ' + stderr.replace(outputReplace, ""));
+						}
+					});
+				} else if (os.type() === 'Linux') { // Penguin check
+					cp.exec(praatPath +'/praat --run ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'/praat --run ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+						let praatOut = vscode.window.createOutputChannel("Script Output (From Praat)");
+						praatOut.show();
+						if (error) {
+							praatOut.appendLine('The script did not run.\nPraat sent:\n' + error);
+						} else {
+							praatOut.appendLine('The script ran nominally.\nPraat sent:\n' + stdout.replace(outputReplace, "") + ' ' + stderr.replace(outputReplace, ""));
+						}
+					});
+				} else if (os.type() === 'Darwin') { // Steve check
+					cp.exec(praatPath +'/Praat.app/Contents/MacOS/Praat --run ' + vscode.window.activeTextEditor?.document.uri.fsPath, (error, stdout, stderr) => {
+						vscode.window.showInformationMessage('Running command: ' + praatPath +'/Praat.app/Contents/MacOS/Praat --run ' + vscode.window.activeTextEditor?.document.uri.fsPath);
+						let praatOut = vscode.window.createOutputChannel("Script Output (From Praat)");
+						praatOut.show();
+						if (error) {
+							praatOut.appendLine('The script did not run.\nPraat sent:\n' + error);
+						} else {
+							praatOut.appendLine('The script ran nominally.\nPraat sent:\n' + stdout.replace(outputReplace, "") + ' ' + stderr.replace(outputReplace, ""));
+						}
+					});
+				} else {
+					// What distro are they even running?
+					vscode.window.showInformationMessage('Fatal error. Trouble recognizing OS.');
+				}
+			} else {
+				vscode.window.showErrorMessage('You do not have a script open on VSCode!');
+			}
+		}
+	});
+
+	// Command for getting user's Praat version
+	registerCommandNice(context, 'praatvscode.getPraatVersion', () => {
+		if (praatPath === undefined) {
+			vscode.window.showErrorMessage('You have not set a path for the Praat executable yet. Run the command "Define a path for the Praat executable" to set a path.');
+		} else {
+			if (os.type() === 'Windows_NT') { // Gate check
+				cp.exec(praatPath +'\\Praat.exe --version', (error, stdout, stderr) => {
+					if (error) {
+						vscode.window.showInformationMessage('Cannot fetch Praat version info.');
+					} else {
+						vscode.window.showInformationMessage('Praat version info: ' + stdout.replace(outputReplace,  ""));
+					}
+				});
+			} else if (os.type() === 'Linux') { // Penguin check
+				cp.exec(praatPath +'/praat --version', (error, stdout, stderr) => {
+					if (error) {
+						vscode.window.showInformationMessage('Cannot fetch Praat version info.');
+					} else {
+						vscode.window.showInformationMessage('Praat version info: ' + stdout.replace(outputReplace, ""));
+				}});
+			} else if (os.type() === 'Darwin') { // Steve check
+				cp.exec(praatPath +'/Praat.app/Contents/MacOS/Praat --version', (error, stdout, stderr) => {
+					if (error) {
+						vscode.window.showInformationMessage('Cannot fetch Praat version info.');
+					} else {
+						vscode.window.showInformationMessage('Praat version info: ' + stdout.replace(outputReplace, ""));
+				}});
+			} else {
+				vscode.window.showInformationMessage('Fatal error. Trouble recognizing OS.');
+			}
 		}
 	});
 }